Extract _gather_members()
